requirements.txt
python-telegram-bot==20.7
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
python-dotenv==1.0.0
girth==0.8.0
matplotlib
seaborn
pytz
PyMuPDF==1.23.8
apscheduler
asyncpg
sqlalchemy[asyncio]
telegram
